// src/models/unit-ontology-term.ts
export interface UnitOntologyTerm {
    classId: string | undefined;
    preferredLabel: string | undefined;
    synonyms: string | undefined;
    definitions: string | undefined;
    obsolete: boolean | undefined;
    cui: string | undefined;
    semanticTypes: string | undefined;
    parents: string | undefined;
    hasRelatedSynonym: string | undefined;
}